requirements.txt
streamlit>=1.31.0
requests>=2.31.0
python-dotenv>=1.0.0 
numpy>=1.24.0
pandas>=2.0.0
plotly>=5.18.0
loguru>=0.7.0
httpx>=0.25.0
orjson>=3.9.0
elommr>=0.1.0
artemis-client>=0.1.0
falcon-models>=0.1.0
vision-models>=0.1.0 